requirements.txt
Django==4.1.4
djangorestframework==3.14.0
psycopg2>=2.8.6
drf-spectacular>=0.15.1
